+++ src/EventBus/Middleware/TearDownAfterEvent.php
@@ -14,7 +14,7 @@ namespace CloudCreativity\Modules\EventBus\Middleware;
 use Closure;
 use CloudCreativity\Modules\Toolkit\Messages\IntegrationEventInterface;
 
-final class TearDownAfterEvent implements IntegrationEventMiddlewareInterface
+final class TearDownAfterEvent implements EventBusMiddlewareInterface
 {
     /**
      * TearDownAfterEvent constructor.